requirements.txt
channels==4.0.0
Django==5.0.1
django-registration==3.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
PyJWT==2.8.0
django-osm-field==0.5.0
requests
django-cors-headers
python-decouple
daphne